separated profile and settings folders
