+++ apps/ui-sharethrift/src/components/layouts/home/account/index.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom';
-import EditSettings from './pages/EditSettings.tsx';
-import Profile from './pages/Profile.tsx';
-import Settings from './pages/Settings.tsx';
+import EditSettings from './settings/pages/EditSettings.tsx';
+import Profile from './profile/pages/Profile.tsx';
+import Settings from './settings/pages/Settings.tsx';
 import AdminDashboardMain from './admin-dashboard/pages/admin-dashboard-main.tsx';
 
 export default function AccountRoutes() {
