--- apps/ui-sharethrift/src/components/layouts/home/account/pages/Main.tsx
@@ -1,3 +1,0 @@
-export default function AccountMain() {
-  return <div>Account Main Page</div>;
-}
