+++ apps/ui-sharethrift/src/components/layouts/home/account/settings/components/settings-view.module.css
