+++ apps/ui-sharethrift/src/components/layouts/home/account/settings/stories/SettingsPage.stories.tsx
@@ -2,8 +2,8 @@ import type { Meta, StoryFn } from "@storybook/react";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
 import { AuthContext } from "react-oidc-context";
 import { type ReactNode, useMemo } from "react";
-import HomeRoutes from "../../index.tsx";
-import { HomeAccountSettingsViewContainerCurrentUserDocument } from "../../../../../generated.tsx";
+import HomeRoutes from "../../../index.tsx";
+import { HomeAccountSettingsViewContainerCurrentUserDocument } from "../../../../../../generated.tsx";
 
 export default {
   title: "Pages/Account/Settings",
